Condensed helpers to one file
